pyproject.toml
[tool.poetry]
name = "django_inertia_svelte_example"
version = "0.1.0"
description = ""
authors = ["Armedi"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^4.2.7"
django-vite = "^3.0.0"
inertia-django = "^0.5.3"
django-js-routes = "^0.2.0"
django-debug-toolbar = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
